[BE] refactor: 랜덤 숫자 생성 클래스 공용 모듈로 이동 #798
