+++ backend/app/src/main/java/com/onair/hearit/app/advertisement/application/AdvertisementService.java
@@ -1,13 +1,11 @@
 package com.onair.hearit.app.advertisement.application;
 
 import com.onair.hearit.app.advertisement.dto.AdvertisementResponse;
+import com.onair.hearit.app.common.RandomNumberGenerator;
 import com.onair.hearit.app.exception.custom.NotFoundException;
 import com.onair.hearit.core.domain.Advertisement;
 import com.onair.hearit.core.infrastructure.jpa.AdvertisementRepository;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.Random;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -19,6 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class AdvertisementService {
 
     private final AdvertisementRepository advertisementRepository;
+    private final RandomNumberGenerator randomNumberGenerator;
 
     @Transactional(readOnly = true)
     public AdvertisementResponse getRandomAdvertisement() {
@@ -27,15 +26,9 @@ public class AdvertisementService {
             log.error("등록된 광고가 존재하지 않습니다.");
             throw new IllegalStateException("등록된 광고가 존재하지 않습니다.");
         }
-        Long randomId = pickRandomId(advertisementIds);
+        Long randomId = advertisementIds.get((int) (randomNumberGenerator.getDouble() * advertisementIds.size()));
         Advertisement advertisement = advertisementRepository.findById(randomId)
                 .orElseThrow(() -> new NotFoundException("advertisementId", randomId.toString()));
         return AdvertisementResponse.from(advertisement);
-    }
-
-    private Long pickRandomId(List<Long> ids) {
-        List<Long> mutableIds = new ArrayList<>(ids);
-        Collections.shuffle(mutableIds, new Random());
-        return mutableIds.get(0);
     }
 }
