+++ backend/app/src/main/java/com/onair/hearit/app/common/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package com.onair.hearit.app.explore.application.scorefactor;
+package com.onair.hearit.app.common;
 
 public interface RandomNumberGenerator {
