+++ backend/app/src/test/java/com/onair/hearit/app/explore/application/scoreprocessor/GuestExploreScoreProcessorTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import com.onair.hearit.app.explore.application.ExploreScoreCalculator;
 import com.onair.hearit.app.explore.application.ExploreScoreInitializer;
 import com.onair.hearit.app.explore.application.scorefactor.BookmarkScoreFactor;
-import com.onair.hearit.app.explore.application.scorefactor.RandomNumberGenerator;
+import com.onair.hearit.app.common.RandomNumberGenerator;
 import com.onair.hearit.app.explore.application.scorefactor.RandomScoreFactor;
 import com.onair.hearit.app.explore.application.scorefactor.RecencyScoreFactor;
 import com.onair.hearit.app.explore.dto.ExploredHearitResponse;
